rename ChartContainer to ChartDataProvider
