--- packages/x-charts/src/ChartContainer/index.ts
@@ -1,1 +1,0 @@
-export * from './ChartContainer';
