+++ packages/x-charts/src/ChartDataProvider/ChartDataProvider.test.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 import { createRenderer, describeConformance } from '@mui/internal-test-utils';
-import { ChartContainer } from '@mui/x-charts/ChartContainer';
+import { ChartDataProvider } from '@mui/x-charts/ChartDataProvider';
 
-describe('<ChartContainer />', () => {
+describe('<ChartDataProvider />', () => {
   const { render } = createRenderer();
 
-  describeConformance(<ChartContainer height={100} width={100} series={[]} />, () => ({
+  describeConformance(<ChartDataProvider height={100} width={100} series={[]} />, () => ({
     classes: {} as any,
     inheritComponent: 'svg',
     render,
-    muiName: 'MuiChartContainer',
+    muiName: 'MuiChartDataProvider',
     testComponentPropWith: 'div',
     refInstanceof: window.SVGSVGElement,
     skip: [
