+++ packages/x-charts/src/ChartDataProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './ChartDataProvider';
